requirements.txt
streamlit==1.37.0
ultralytics==8.0.163
Pillow==10.0.0
firebase-admin==6.5.0
bcrypt==4.1.3
